requirements.txt
Django==1.7
PyJWT==0.3.0
Pygments==1.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.17
django-model-utils==2.2
django-websocket-redis==0.4.4
djangorestframework==3.1.2
gevent==1.0.1
greenlet==0.4.6
ipython==2.3.0
kombu==3.0.24
oauthlib==0.7.1
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2015.2
redis==2.10.3
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
wsgiref==0.1.2
